requirements.txt
numpy==2.0.2
matplotlib==3.9.4
imageio==2.37.0
scipy==1.13.1
trimesh==4.6.2
pyyaml==6.0.2
cachetools==5.5.1
CacheToolsUtils==6.0
networkx==3.2.1
rtree==1.3.0
mujoco==3.3.0
gymnasium==0.28.1
stable-baselines3==2.3.2
opencv-python-headless==4.11.0.86
torch==2.6.0
scikit-image==0.24.0